+++ apps/actors/main.tsx
@@ -10,7 +10,7 @@ import useHashParam from "use-hash-param";
 import { SystemInstance, SystemInstanceAction } from "./types";
 import { useEffectfulReducer } from "../../uiCommon/generic/hooks";
 import { CollapsibleWithHeading } from "../../uiCommon/generic/collapsible";
-import { Window } from "./window";
+import { MultiClient } from "./ui/multiClient";
 
 const initialSystemsState = initialState(SYSTEMS);
 
@@ -75,70 +75,6 @@ function SystemInstanceView<St extends Json, Msg extends Json>(props: {
         displayDataTypes={false}
         collapsed
       />
-    </>
-  );
-}
-
-function MultiClient<St extends Json, Msg extends Json>(props: {
-  systemInstance: SystemInstance<St, Msg>;
-  // hoo that is a big type
-  dispatch: (action: SystemInstanceAction<St, Msg>) => void;
-}) {
-  const sendInput = (clientID: string, input: Msg) => {
-    props.dispatch({
-      type: "UpdateTrace",
-      action: {
-        type: "SendUserInput",
-        clientID,
-        input,
-      },
-    });
-  };
-
-  return (
-    <>
-      <button
-        onClick={() => {
-          props.dispatch({ type: "AllocateClientID" });
-          // TODO: DRY this up with other place client id is constructed
-          const clientID = `client${props.systemInstance.nextClientID}`;
-          props.dispatch({
-            type: "UpdateTrace",
-            action: {
-              type: "SpawnClient",
-              id: props.systemInstance.nextClientID.toString(),
-              initialUserState: props.systemInstance.system.initialUserState,
-              initialClientState:
-                props.systemInstance.system.initialClientState(clientID),
-            },
-          });
-        }}
-      >
-        Add Client
-      </button>
-      <div style={{ display: "flex", flexDirection: "row" }}>
-        {props.systemInstance.clientIDs.map((clientID) => {
-          const clientState =
-            props.systemInstance.trace.latestStates[`client${clientID}`];
-
-          return (
-            <Window
-              key={clientID}
-              name={`Client ${clientID}`}
-              onClose={() => {
-                props.dispatch({ type: "ExitClient", clientID });
-              }}
-            >
-              {clientState ? (
-                <props.systemInstance.system.ui
-                  state={clientState}
-                  sendUserInput={(input) => sendInput(clientID, input)}
-                />
-              ) : null}
-            </Window>
-          );
-        })}
-      </div>
     </>
   );
 }
